Code fix after merge

+++ src/resqml2_2/SealedSurfaceFrameworkRepresentation.cpp
@@ -133,9 +133,9 @@ void SealedSurfaceFrameworkRepresentation::pushBackContactPatch(
     resqml22__SealedContact* contactRep = static_cast<resqml22__SealedContact*>(orgRep->Contacts[contactIndex]);
 
     // we look for the supporting representation index
-    int representationIndex = -1;
-	const unsigned int representationCount = getRepresentationCount();
-	for (unsigned int i = 0; i < representationCount; ++i) {
+	int64_t representationIndex = -1;
+	const uint64_t representationCount = getRepresentationCount();
+	for (uint64_t i = 0; i < representationCount; ++i) {
 		if (getRepresentation(i)->getUuid() == supportingRepresentation->getUuid()) {
 			representationIndex = i;
 			break;
